refactor and add mobile view
